requirements.txt
robotframework>=6.0
robotframework-requests>=0.9.0
python-dotenv>=1.0.0 